Rename css files to scss and install jest-dom
