+++ package.json
@@ -48,8 +48,6 @@
     "@openmrs/esm-extensions": "^3.0.0",
     "@openmrs/esm-config": "^3.0.0",
     "@openmrs/esm-styleguide": "^3.0.0",
-    "carbon-components": "^10.19.0",
-    "carbon-icons": "^7.0.7",
     "i18next": "^19.7.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -87,9 +85,6 @@
     "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.6",
     "browserslist-config-openmrs": "^1.0.0",
-    "carbon-components": "^10.24.0",
-    "carbon-components-react": "^7.24.0",
-    "carbon-icons": "^7.0.7",
     "clean-webpack-plugin": "^3.0.0",
     "concurrently": "^5.0.0",
     "css-loader": "^3.2.0",
