+++ src/patient-registration/input/custom-input/autosuggest/autosuggest.scss
