+++ src/patient-registration/input/custom-input/autosuggest/autosuggest.test.tsx
@@ -3,6 +3,7 @@ import { Autosuggest } from './autosuggest.component';
 import { render, screen, fireEvent, waitForElement } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import '@testing-library/jest-dom/extend-expect';
+import '@testing-library/jest-dom';
 
 const mockPersons = [
   {
